refactor(core): restructure core module and services for better maintainability
This commit introduces a new `core` module with services, exceptions, filters, guards, and pipes. It also refactors the `DbMain` class for improved readability and error handling. The user authentication and registration logic have been centralized in the `UserService` and `TokenService` for better reusability and maintainability. Additionally, the `Cripto` and `PwdEncrypt` classes have been enhanced with static methods for encryption and decryption. The commit also includes the addition of rate-limiting and exception handling mechanisms to improve security and error management.